feat(diri): move type conversion to orderbook/events
