+++ crates/diri/src/orderbook/events/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod common;
 pub(crate) mod order_cancelled;
 pub(crate) mod order_executed;
 pub(crate) mod order_fulfilled;
@@ -10,7 +11,7 @@ pub(crate) use order_fulfilled::OrderFulfilled;
 pub(crate) use order_placed::OrderPlaced;
 pub(crate) use rollback_status::RollbackStatus;
 
-pub(crate) use cainome::cairo_serde::U256;
+// pub(crate) use cainome::cairo_serde::U256;
 use starknet::{core::types::Felt, macros::selector};
 
 #[derive(Debug)]
